rag/requirements-rag.txt
chromadb==0.4.24
sentence-transformers==2.7.0
pypdf==3.17.4
httpx==0.25.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
